requirements.txt
aiofiles==0.7.0
aioredis==2.0.0
alembic==1.6.5
asyncpg==0.24.0
databases[postgresql]==0.5.1
fastapi==0.68.0
itsdangerous==2.0.1
psycopg2==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
SQLAlchemy==1.4.22
uvicorn==0.14.0

